reorg steps into separate files for easier editing
